+++ specula/processing_objects/atmo_random_phase.py
@@ -2,7 +2,7 @@ import numpy as np
 from astropy.io import fits
 
 from specula.base_processing_obj import BaseProcessingObj
-from specula.data_objects.ef import ElectricField
+from specula.data_objects.electric_field import ElectricField
 from specula.base_value import BaseValue
 from specula.data_objects.layer import Layer
 from specula.data_objects.pupilstop import Pupilstop
