+++ specula/processing_objects/modulated_pyramid.py
@@ -3,7 +3,7 @@ from specula import fuse, show_in_profiler
 from specula.base_processing_obj import BaseProcessingObj
 from specula.base_value import BaseValue
 from specula.connections import InputValue
-from specula.data_objects.ef import ElectricField
+from specula.data_objects.electric_field import ElectricField
 from specula.lib.make_xy import make_xy
 from specula.data_objects.intensity import Intensity
 from specula.lib.make_mask import make_mask
